# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ChangePasswordRequest": {
        "additionalProperties": false,
        "properties": {
          "current_password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Current Password",
            "type": "string"
          },
          "new_password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "New Password",
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "ChangePasswordRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InviteAcceptRequest": {
        "additionalProperties": false,
        "properties": {
          "company_code": {
            "maxLength": 63,
            "minLength": 1,
            "title": "Company Code",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "token": {
            "maxLength": 128,
            "minLength": 16,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "company_code",
          "token",
          "email",
          "password"
        ],
        "title": "InviteAcceptRequest",
        "type": "object"
      },
      "InviteCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "email_domain": {
            "anyOf": [
              {
                "maxLength": 253,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email Domain"
          },
          "max_uses": {
            "anyOf": [
              {
                "maximum": 1000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Uses"
          },
          "ttl_days": {
            "default": 7,
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Ttl Days",
            "type": "integer"
          }
        },
        "title": "InviteCreateRequest",
        "type": "object"
      },
      "InviteCreatedResponse": {
        "properties": {
          "company_code": {
            "title": "Company Code",
            "type": "string"
          },
          "invite": {
            "$ref": "#/components/schemas/InviteResponse"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "invite",
          "token",
          "company_code"
        ],
        "title": "InviteCreatedResponse",
        "type": "object"
      },
      "InvitePreviewResponse": {
        "properties": {
          "company_name": {
            "title": "Company Name",
            "type": "string"
          },
          "email_domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email Domain"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          }
        },
        "required": [
          "company_name",
          "expires_at",
          "email_domain"
        ],
        "title": "InvitePreviewResponse",
        "type": "object"
      },
      "InviteResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email_domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email Domain"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "max_uses": {
            "title": "Max Uses",
            "type": "integer"
          },
          "status": {
            "enum": [
              "active",
              "expired",
              "revoked",
              "used_up"
            ],
            "title": "Status",
            "type": "string"
          },
          "uses": {
            "title": "Uses",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "created_at",
          "expires_at",
          "max_uses",
          "uses",
          "email_domain",
          "status"
        ],
        "title": "InviteResponse",
        "type": "object"
      },
      "InviteTokenRequest": {
        "additionalProperties": false,
        "properties": {
          "company_code": {
            "maxLength": 63,
            "minLength": 1,
            "title": "Company Code",
            "type": "string"
          },
          "token": {
            "maxLength": 128,
            "minLength": 16,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "company_code",
          "token"
        ],
        "title": "InviteTokenRequest",
        "type": "object"
      },
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "company_code": {
            "maxLength": 63,
            "minLength": 1,
            "title": "Company Code",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "company_code",
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "MeResponse": {
        "properties": {
          "company_code": {
            "title": "Company Code",
            "type": "string"
          },
          "company_name": {
            "title": "Company Name",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_login_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "must_change_password": {
            "title": "Must Change Password",
            "type": "boolean"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "tenant_id",
          "company_name",
          "company_code",
          "must_change_password",
          "last_login_at"
        ],
        "title": "MeResponse",
        "type": "object"
      },
...
      "RefreshRequest": {
        "additionalProperties": false,
        "properties": {
          "refresh_token": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "expires_in"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
      "UserRole": {
        "description": "Роль сотрудника внутри своей компании.\n\nADMIN — управляет документами и пользователями компании, видит\nотладку поиска и отчёт о пробелах. EMPLOYEE — задаёт вопросы.\nЗаводить компании (тенанты) не может ни одна роль: это делает\nкоманда Kronto через CLI на сервере (см. corp_ed.cli).",
        "enum": [
          "admin",
          "employee"
        ],
        "title": "UserRole",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Kronto — ИИ-ассистент по внутренним документам компании",
    "title": "corp-ed",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/change-password": {
      "post": {
        "description": "Сменить пароль. Все прежние сессии закрываются, возвращается\nновая пара токенов для текущего устройства.\n\nЛимит — против подбора текущего пароля украденным access-токеном.",
        "operationId": "change_password_api_v1_auth_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change Password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "Вход. Два лимита против перебора (ASVS 6.3.1):\n\n- на IP — все попытки: один адрес не перебирает много учёток;\n- на учётку — только неудачные: распределённый перебор одной учётки\n  с многих адресов. Успешный вход счётчик обнуляет.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout-all": {
      "post": {
        "operationId": "logout_everywhere_api_v1_auth_logout_all_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Logout Everywhere",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "Кто вошёл. Доступна и до смены временного пароля: фронту нужно\nзнать must_change_password, чтобы показать форму смены.",
        "operationId": "read_me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Me",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/invites": {
      "get": {
        "description": "Последние ссылки компании с их состоянием; токенов в ответе нет.",
        "operationId": "list_invites_api_v1_invites_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InviteResponse"
                  },
                  "title": "Response List Invites Api V1 Invites Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Invites",
        "tags": [
          "invites"
        ]
      },
      "post": {
        "description": "Ссылка-приглашение в свою компанию. Токен — в ответе один раз.",
        "operationId": "create_invite_api_v1_invites_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteCreatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Invite",
        "tags": [
          "invites"
        ]
      }
    },
    "/api/v1/invites/accept": {
      "post": {
        "description": "Завести учётку сотрудника по ссылке и сразу войти.",
        "operationId": "accept_invite_api_v1_invites_accept_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteAcceptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Invite",
        "tags": [
          "invites"
        ]
      }
    },
    "/api/v1/invites/preview": {
      "post": {
        "description": "В какую компанию ведёт ссылка — для карточки «Присоединиться».",
        "operationId": "preview_invite_api_v1_invites_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitePreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Invite",
        "tags": [
          "invites"
        ]
      }
    },
    "/api/v1/invites/{invite_id}": {
      "delete": {
        "operationId": "revoke_invite_api_v1_invites__invite_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "invite_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invite Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Invite",
        "tags": [
          "invites"
        ]
      }
    },
...
  }
}